moved figma colours to theme
