+++ Sources/W3WSwiftThemes/Design/Schemes/Colors/Enums/W3WColorMode.swift
