+++ Sources/W3WSwiftThemes/Design/Schemes/Styles/StyleTypes/W3WIconSize.swift
