+++ Sources/W3WSwiftThemes/Design/Themes/W3WTheme.swift
@@ -14,6 +14,53 @@ public class W3WTheme: CustomDebugStringConvertible, CustomStringConvertible {
   // this holds a dictionary of schemes.  schemes contain colours and styles for a particular type of element such as textFeilds, or buttons, etc
   public var schemes = [W3WSetTypes : W3WScheme]()
 
+  // MARK: Vars
+  
+  public var brandBase: W3WColor?
+  
+  public var labelsPrimary: W3WColor?
+  public var labelsSecondary: W3WColor?
+  public var labelsTertiary: W3WColor?
+  public var labelsQuaternary: W3WColor?
+  public var labelsPrimaryInverse: W3WColor?
+  public var labelsPrimaryBlackInverse: W3WColor?
+
+  public var fillsPrimary: W3WColor?
+  public var fillsSecondary: W3WColor?
+  public var fillsTertiary: W3WColor?
+  public var fillsQuaternary: W3WColor?
+
+  public var separatorOpaque: W3WColor?
+  public var separatorNonOpaque: W3WColor?
+  
+  public var systemBackgroundBasePrimary: W3WColor?
+  public var systemBackgroundBaseSecondary: W3WColor?
+  public var systemBackgroundBaseTertiary: W3WColor?
+
+  public var systemBackgroundElevatedPrimary: W3WColor?
+  public var systemBackgroundElevatedSecondary: W3WColor?
+  public var systemBackgroundElevatedTertiary: W3WColor?
+
+  public var groupedBackgroundBasePrimary: W3WColor?
+  public var groupedBackgroundBaseSecondary: W3WColor?
+  public var groupedBackgroundBaseTertiary: W3WColor?
+
+  public var groupedBackgroundElevatedPrimary: W3WColor?
+  public var groupedBackgroundElevatedSecondary: W3WColor?
+  public var groupedBackgroundElevatedTertiary: W3WColor?
+
+  public var successBase: W3WColor?
+  public var successElevated: W3WColor?
+  public var successLabel: W3WColor?
+
+  public var warningBase: W3WColor?
+  public var warningElevated: W3WColor?
+  public var warningLabel: W3WColor?
+
+  public var errorBase: W3WColor?
+  public var errorElevated: W3WColor?
+  public var errorLabel: W3WColor?
+
   
   /// Init
 
@@ -45,6 +92,79 @@ public class W3WTheme: CustomDebugStringConvertible, CustomStringConvertible {
     //self.styles = theme?.styles
   }
   
+  
+  /// initialise with a foreground, background, highlight and secondary colour
+  public init(brandBase: W3WColor? = nil, labelsPrimary: W3WColor? = nil, labelsSecondary: W3WColor? = nil, 
+              labelsTertiary: W3WColor? = nil, labelsQuaternary: W3WColor? = nil, labelsPrimaryInverse: W3WColor? = nil,
+              labelsPrimaryBlackInverse: W3WColor? = nil, fillsPrimary: W3WColor? = nil, fillsSecondary: W3WColor? = nil,
+              fillsTertiary: W3WColor? = nil, fillsQuaternary: W3WColor? = nil, separatorOpaque: W3WColor? = nil,
+              separatorNonOpaque: W3WColor? = nil, systemBackgroundBasePrimary: W3WColor? = nil,
+              systemBackgroundBaseSecondary: W3WColor? = nil, systemBackgroundBaseTertiary: W3WColor? = nil,
+              systemBackgroundElevatedPrimary: W3WColor? = nil, systemBackgroundElevatedSecondary: W3WColor? = nil,
+              systemBackgroundElevatedTertiary: W3WColor? = nil, groupedBackgroundBasePrimary: W3WColor? = nil,
+              groupedBackgroundBaseSecondary: W3WColor? = nil, groupedBackgroundBaseTertiary: W3WColor? = nil,
+              groupedBackgroundElevatedPrimary: W3WColor? = nil, groupedBackgroundElevatedSecondary: W3WColor? = nil,
+              groupedBackgroundElevatedTertiary: W3WColor? = nil, successBase: W3WColor? = nil, successElevated: W3WColor? = nil,
+              successLabel: W3WColor? = nil, warningBase: W3WColor? = nil, warningElevated: W3WColor? = nil, warningLabel: W3WColor? = nil,
+              errorBase: W3WColor? = nil, errorElevated: W3WColor? = nil, errorLabel: W3WColor? = nil,
+              base: W3WScheme = .w3w, buttons: W3WScheme? = nil, textFields: W3WScheme? = nil, labels: W3WScheme? = nil, maps: W3WScheme? = nil,
+              cells: W3WScheme? = nil, icons: W3WScheme? = nil, voice: W3WScheme? = nil, ocr: W3WScheme? = nil) {
+    self.brandBase = brandBase
+
+    self.labelsPrimary = labelsPrimary
+    self.labelsSecondary = labelsSecondary
+    self.labelsTertiary = labelsTertiary
+    self.labelsQuaternary = labelsQuaternary
+    self.labelsPrimaryInverse = labelsPrimaryInverse
+    self.labelsPrimaryBlackInverse = labelsPrimaryBlackInverse
+
+    self.fillsPrimary = fillsPrimary
+    self.fillsSecondary = fillsSecondary
+    self.fillsTertiary = fillsTertiary
+    self.fillsQuaternary = fillsQuaternary
+
+    self.separatorOpaque = separatorOpaque
+    self.separatorNonOpaque = separatorNonOpaque
+    
+    self.systemBackgroundBasePrimary = systemBackgroundBasePrimary
+    self.systemBackgroundBaseSecondary = systemBackgroundBaseSecondary
+    self.systemBackgroundBaseTertiary = systemBackgroundBaseTertiary
+
+    self.systemBackgroundElevatedPrimary = systemBackgroundElevatedPrimary
+    self.systemBackgroundElevatedSecondary = systemBackgroundElevatedSecondary
+    self.systemBackgroundElevatedTertiary = systemBackgroundElevatedTertiary
+
+    self.groupedBackgroundBasePrimary = groupedBackgroundBasePrimary
+    self.groupedBackgroundBaseSecondary = groupedBackgroundBaseSecondary
+    self.groupedBackgroundBaseTertiary = groupedBackgroundBaseTertiary
+
+    self.groupedBackgroundElevatedPrimary = groupedBackgroundElevatedPrimary
+    self.groupedBackgroundElevatedSecondary = groupedBackgroundElevatedSecondary
+    self.groupedBackgroundElevatedTertiary = groupedBackgroundElevatedTertiary
+
+    self.successBase = successBase
+    self.successElevated = successElevated
+    self.successLabel = successLabel
+
+    self.warningBase = warningBase
+    self.warningElevated = warningElevated
+    self.warningLabel = warningLabel
+
+    self.errorBase = errorBase
+    self.errorElevated = errorElevated
+    self.errorLabel = errorLabel
+
+    schemes[.base] = base
+    add(type: .buttons,    scheme: buttons)
+    add(type: .textFields, scheme: textFields)
+    add(type: .labels,    scheme: labels)
+    add(type: .maps,     scheme: maps)
+    add(type: .cells,   scheme: cells)
+    add(type: .icons,  scheme: icons)
+    add(type: .voice, scheme: voice)
+    add(type: .ocr,  scheme: ocr)
+  }
+
 
   /// copy a color set from one set to another
   public func copy(from: W3WSetTypes, to: W3WSetTypes) -> W3WTheme {
